\name{Activity}
\alias{Activity}
\title{
Activity Constructor
}
\description{
Constructor for Activity
}
\usage{
Activity()
}
\arguments{
  \item{entity}{
(optional) a named list whose elements will be set as the properties of the activity. The list of
valid properties is controlled by Synapse and currently no client-side verification is performed.
}
}

\value{
The created Activity is returned.
}
\author{
Bruce Hoff
}
