\name{addDir-methods}
\docType{methods}
\alias{addDir-methods}
\alias{addDir,Layer,character-method}
\alias{addDir,Layer,missing-method}
\title{ Methods for Function \code{addDir}  }
\description{
	Methods for adding a directory containing files to an entity.
}
\section{Methods}{
\describe{
	
	\item{\code{signature(entity = "Layer", path = "character")}}{
	%%  ~~describe this method here~~
}

\item{\code{signature(entity = "Layer", path = "missing")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}

