\name{addFile-methods}
\docType{methods}
\alias{addFile-methods}
\alias{addFile,CachedLocation,character,character-method}
\alias{addFile,LocationOwner,character,missing-method}
\alias{addFile,LocationOwner,character,character-method}
\title{Methods for Function \code{addFile}}
\description{
  Methods for function \code{addFile}
}
\section{Methods}{
  \describe{
     \item{\code{signature(entity = "File", file = "character", path = "character")}}{
      Set the physical file associated with the File object to be that given by <path,file>.   Replaces
      any other file or object associated with the file.
    }
  }
}
\keyword{methods}
