\name{as.environment}
\alias{as.environment}
\docType{data}
\title{
as.environment
}
\description{
variation of as.environment() for certain S-4 objects in the synapseClient package
}
\usage{
as.environment(enhancedEnvironment)
as.environment(globalCache)
as.environment(archiveOwner)
}
\details{
This function is deprecated and will be removed.  Do not use.
}

