\name{as.list}
\alias{as.list}
\docType{methods}
\title{
as.list
}
\description{
variation of as.list() for certain S-4 objects in the synapseClient package
}
\usage{
as.list(SimplePropertyOwner)
as.list(SynapseAnnotations)
as.list(TypedPropertyStore)
as.list(SynapseProperties)
}
\details{
This function is deprecated and will be removed.  Do not use.
}

