\name{copyEntity-methods}
\docType{methods}
\alias{copyEntity-methods}
\alias{copyEntity,SynapseLocationOwner-method}
\title{Methods for Function \code{copyEntity}}
\description{
Methods for function \code{copyEntity}
}
\section{Methods}{
\describe{

\item{\code{signature(entity = "SynapseLocationOwner")}}{
Make a copy of a SynapseLocationOwner
}
}}
\keyword{methods}
