\name{Dataset}
\alias{Dataset}
\alias{Media}
\alias{Code}
\alias{SynapseEntity}
\alias{Layer}
\alias{Project}
\alias{Analysis}
\alias{Step}
\title{
Entity Constructors
}
\description{
Constructors for Synapse Entities
}
\usage{
SynapseEntity(entity)
Dataset(entity)
Layer(entity)
Project(entity)
}
\arguments{
  \item{entity}{
(optional) a named list whose elements will be set as the properties of the entity
}
}

\value{
The entity
}
\author{
Matt Furia
}
