\name{generatedBy}
\alias{generatedBy}
\title{
Retrieve the Activity which generated an Entity
}
\description{
Function for retrieving the Activity which generated an Entity
}
\usage{
generatedBy(entity)
}
\arguments{
  \item{entity}{
The entity of interest
}
}
\details{
Returns information already in memory.  (No server interaction takes place.)
}
\value{
The Activity which generated the Entity, or NULL if there is no activity.
}

\author{
Bruce Hoff
}

\seealso{
}

