\name{generatedBy<--methods}
\docType{methods}
\alias{generatedBy<--methods}
\alias{generatedBy<-,SynapseEntity,Activity}
\title{Methods for Function \code{generatedBy<-}}
\description{
 	Methods for function \code{generatedBy<-}
}
\section{Methods}{
\describe{

\item{\code{signature(object = "Activity")}}{
Sets the Activity which generated an entity.  NULL is allowed.
}
}}
\keyword{methods}
