\name{getClientConfigEnvironment}
\alias{getClientConfigEnvironment}
\title{
	Get Client Config Environment
}
\description{
	Returns the environment where the client config object will be cached.
}
\usage{
	getClientConfigEnvironment()
}

\details{
	
}
\value{
	Currently returns R_GlobalEnv.
}
\author{
	Matt Furia
}
\note{
	This function is for internal use only.
}
\keyword{ utilities }
\keyword{ environment }
