\name{getClientConfigName}
\alias{getClientConfigName}
\title{
	Get the Name of the Client Config Object
}
\description{
	Returns the name of the object where the client configuration is cached.
}
\usage{
	getClientConfigName()
}
\value{
	Returns the string ".client.config"
}
\author{
	Matt Furia
}
\note{
	This function is for internal use only.
}
\keyword{ utilities }