\name{getLocations}
\alias{getLocations}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	Get Locations for an Object
}
\description{
	Get a list of locations for the data files for the provided object
}
\usage{
	getLocations(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
		The object whose locations should be returned
	}
}
\value{
	List of locations
}
\author{
	Matt Furia
}

\seealso{
	\code{\link{getDatasetLayers}}
}
\examples{
	## Not Run

	## End Not Run
}

