\name{getPackageName}
\alias{getPackageName}
\docType{data}
\title{
getPackageName
}
\description{
variation of getPackageName() for certain S-4 objects in the synapseClient package
}
\usage{
getPackageName(EnhancedEnvironment)
getPackageName(ArchiveOwner)
}
\details{
This function is deprecated and will be removed.  Do not use.
}

