\name{Layer-methods}
\docType{methods}
\alias{Layer-methods}
\alias{Layer,character-method}
\alias{Layer,list-method}
\alias{Layer,numeric-method}
\title{Layer}
\description{
Layer entity constructors.
}
\section{Methods}{
\describe{

\item{\code{signature(entity = "list")}}{
Build a Layer entity from a list of properties.
}

}}
\keyword{methods}
