\name{names}
\alias{names}
\docType{data}
\title{
names
}
\description{
variation of names() for certain S-4 objects in the synapseClient package
}
\usage{
names(CachingObjectOwner)
names(EnhancedEnvironment)
names(LocationableWithoutBinaries)
names(Locationable)
names(Entity)
names(QueryResult)
names(SynapseProperties)
}
\details{
This function is deprecated and will be removed.  Do not use.
}

