\name{PaginatedResults}
\alias{PaginatedResults}
\docType{data}
\title{
PaginatedResults Constructor
}
\description{
Constructor for PaginatedResults
}
\details{
This objects holds a list of objects, "results" and an integer "totalNumberofResults", which 
is the total number of objects across all pages.
}
\usage{new("PaginatedResults")}
\seealso{
\code{\link{synGetSubmissions}}
\code{\link{synGetOwnSubmissions}}
}

