\name{properties}
\alias{properties}
\alias{propertyValue<-}
\alias{propertyValue}
\alias{propertyValues<-}
\alias{propertyValues}
\alias{deleteProperty}
\title{
Functions for Manipulating Entity Properties
}
\description{
Manage object properties
}
\usage{
properties(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
The Synapse object of interest
}
}
