\name{setPackageName}
\alias{setPackageName}
\docType{methods}
\title{
setPackageName
}
\description{
variation of setPackageName() for certain S-4 objects in the synapseClient package
}
\usage{
setPackageName(where, create=TRUE)
}
\details{
This function is deprecated and will be removed.  Do not use.
}

