\name{synapseDataLocationPreferences}
\alias{synapseDataLocationPreferences}
\title{
	Data Location Preferences
}
\description{
	Set and retrieve the user's data location preferences.
}
\usage{
	synapseDataLocationPreferences(locationTypes)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{locationTypes}{
		an ordered array of locations types that will be used to get data from Synapse
	}
}
\value{
	If no arguments are provided, returns the array of locations from the memory cache.
}

\author{
	Matt Furia
}
