\name{SynapseEntityMethods}
\alias{addFile-methods}
\alias{addFile,CachedLocation,character,character-method}
\alias{addFile,LocationOwner,character,missing-method}
\alias{addFile,LocationOwner,character,character-method}
\alias{addFile}
\alias{deleteFile}
\alias{moveFile}
\alias{addObject-methods}
\alias{addObject,SynapseLocationOwner,ANY-method}
\alias{addObject,SynapseLocationOwner,data.frame-method}
\alias{addObject,SynapseLocationOwner,list-method}
\alias{addObject}
\alias{deleteObject}
\alias{getObject}
\alias{listObjects}
\alias{renameObject}
\alias{getEntity}
\alias{getEntity-methods}
\alias{getEntity,character-method}
\alias{getEntity,list-method}
\alias{getEntity,numeric-method}
\alias{getEntity,SynapseEntity-method}
\alias{deleteFile-methods}
\alias{deleteFile,FileCache,character-method}
\alias{deleteFile,SynapseLocationOwner,character-method}
\alias{deleteEntity-methods}
\alias{deleteEntity,character-method}
\alias{deleteEntity,list-method}
\alias{deleteEntity,numeric-method}
\alias{deleteEntity,SynapseEntity-method}
\alias{deleteObject-methods}
\alias{deleteObject,SynapseLocationOwner,character-method}
\alias{moveFile-methods}
\alias{moveFile,FileCache-method}
\alias{moveFile,SynapseLocationOwner-method}
\alias{getObject-methods}
\alias{getObject,SynapseLocationOwner,character-method}
\alias{renameObject-methods}
\alias{renameObject,SynapseLocationOwner,character,character-method}
\alias{storeEntity-methods}
\alias{storeEntity,SynapseLocationOwner-method}
\alias{storeEntity,SynapseEntity-method}
\alias{createEntity}
\alias{getEntity}
\alias{deleteEntity}
\alias{updateEntity}
\alias{storeEntity}
\alias{storeEntityFiles}
\alias{storeEntityObjects}
\alias{downloadEntity}
\alias{loadEntity}
\alias{getParentEntity}
\alias{createEntity-methods}
\alias{createEntity,SynapseLocationOwner-method}
\alias{createEntity,SynapseEntity-method}
\alias{updateEntity-methods}
\alias{updateEntity,SynapseLocationOwner-method}
\alias{updateEntity,SynapseEntity-method}
\alias{downloadEntity-methods}
\alias{downloadEntity,character-method}
\alias{downloadEntity,SynapseLocationOwner-method}
\alias{downloadEntity,numeric-method}
\alias{downloadEntity,SynapseEntity-method}
\alias{loadEntity-methods}
\alias{loadEntity,character-method}
\alias{loadEntity,SynapseLocationOwner-method}
\alias{loadEntity,numeric-method}


\title{
DEPRECATED Entity Methods
}
\description{
Methods specific to deprecated Synapse Entities. These methods are now also deprecated. Please see \code{\link{Project}}, \code{\link{Folder}}, \code{\link{File}} for currently supported Synapse Objects and \code{\link{synGet}}, \code{\link{synStore}} for most commonly used methods.
}
\usage{
DEPRECATED
}
\arguments{
DEPRECATED
}

\value{
DEPRECATED
}
