\name{synapseGet}
\alias{synapseGet}
\alias{synapsePost}
\alias{synapseDelete}
\alias{synapsePut}
\alias{.synapsePostPut}
\alias{.synapseGetDelete}
\title{
	Synapse REST
}
\description{
	Functions providing a low-level interface to the Synapse REST API
}
\usage{
	synapseGet(uri, host = .getRepoEndpointLocation(), curlHandle = getCurlHandle(), anonymous = .getCache("anonymous"), path = .getRepoEndpointPrefix(), opts = .getCache("curlOpts"))
	synapseDelete(uri, entity, host = .getRepoEndpointLocation(), curlHandle = getCurlHandle(), anonymous = FALSE, path = .getRepoEndpointPrefix(), opts = .getCache("curlOpts"))
	synapsePut(uri, entity, host = .getRepoEndpointLocation(), curlHandle = getCurlHandle(), anonymous = FALSE, path = .getRepoEndpointPrefix(), opts = .getCache("curlOpts"))
	synapsePost(uri, entity, host = .getRepoEndpointLocation(), curlHandle = getCurlHandle(), anonymous = FALSE, path = .getRepoEndpointPrefix(), opts = .getCache("curlOpts"))
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{uri}{
  		uri of the request
	}
  \item{entity}{
  		entity to be put or posted
	}
  \item{host}{
  		hostname of the server
	}
  \item{curlHandle}{
  		curlHandle to be used for the request
	}
  \item{anonymous}{
  		if TRUE, the request will be made anonymously
	}
  \item{path}{
  		path to the service
	}
  \item{opts}{
  		curl options
	}
}

\value{
	an R list generated by calling fromJSON on the server response.
}
\author{
	Matt Furia
}

\seealso{
	\code{\link{curlPerform}}
}

