\name{synDownloadTableColumns}
\alias{synDownloadTableColumns}
\docType{methods}
\title{
synDownloadTableColumns
}
\description{
Create Table Columns based on a data frame
}
\usage{
  synDownloadTableColumns(synTable, tableColumns)
}
\arguments{
  \item{synTable}{
	 TODO
  }
  \item{tableColumns}{
  	TODO
  }
}
\value{
TODO
}

