\name{synSetActivity}
\alias{synSetActivity}
\title{
\code{synSetActivity}
}
\description{
	Function for setting the Activity that generates an entity
}
\usage{
	synSetActivity(entity)<-activity
}
\arguments{
  \item{entity}{
	the entity generated by the Activity
}
}
\details{
	Sets the Activity that generates a given entity
}
\value{
	an Entity object
}

\author{
Bruce Hoff
}
