\name{touOnWeb}
\alias{touOnWeb}
\title{
View the Synapse Terms of Use as a Web Page 
}
\description{
Show the Synapse Terms of Use as a Web Page
}
\usage{
touOnWeb()
}
\details{
Opens the default web browser to the Synapse Terms of Use web page. These are the same general Terms of Use you must agree to, before being allowed to log in to Synapse.
}

\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
