\name{WikiHeader}
\alias{WikiHeader}
\docType{methods}
\title{
WikiHeader Constructor
}
\usage{WikiHeader(id, title, parentId)}
\arguments{
  \item{id}{
  The ID of this wiki page. (string)
  }
  \item{title}{
  The title to of this page. (string)
  }
  \item{parentId}{
  The parentID to of this page. (string)
  }
}
\section{Slots}{
  \describe{
    \item{\code{id}}{
    The ID of this wiki page. (string)
    }
    \item{\code{title}}{
    The title to of this page. (string)
    }
    \item{\code{parentId}}{
    The parentID to of this page. (string)
    }
  }
}
\seealso{
}