\name{addFile-methods}
\docType{methods}
\alias{addFile-methods}
\alias{addFile,CachedLocation,character,character-method}
\alias{addFile,Layer,character,ANY-method}
\alias{addFile,Layer,character,character-method}
\alias{addFile,Layer,missing,character-method}
\alias{addFile,Layer,missing,missing-method}
\title{Methods for Function \code{addFile}}
\description{
	Methods for function \code{addFile}
}
\section{Methods}{
\describe{

\item{\code{signature(entity = "CachedLocation", file = "character", path = "character")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(entity = "Layer", file = "character", path = "ANY")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(entity = "Layer", file = "character", path = "character")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(entity = "Layer", file = "missing", path = "character")}}{
%%  ~~describe this method here~~
}

\item{\code{signature(entity = "Layer", file = "missing", path = "missing")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
