\name{copyEntity-methods}
\docType{methods}
\alias{copyEntity-methods}
\alias{copyEntity,Layer-method}
\title{Methods for Function \code{copyEntity}}
\description{
	Methods for function \code{copyEntity}
}
\section{Methods}{
\describe{

\item{\code{signature(entity = "Layer")}}{
	Make a copy of a layer
}
}}
\keyword{methods}
