\name{synapseClient-package}
\alias{synapseClient-package}
\alias{synapseClient}
\docType{package}
\title{
Synapse Client
}
\description{
A client for interacting with the Synapse web services.
}
\details{
\tabular{ll}{
Package: \tab synapseClient\cr
Type: \tab Package\cr
Version: \tab 1.1-0\cr
Date: \tab 2014-01-17\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
Matt Furia and Nicole Deflaux
Maintainer: Bruce Hoff <bruce.hoff@sagebase.org>, Chris Bare <chris.bare@sagebase.org>
}
\keyword{ package }
\examples{
## Not Run

## End Not Run
}
