\name{synapseClient-package}
\alias{synapseClient-package}
\alias{synapseClient}
\docType{package}
\title{
Synapse Client
}
\description{
A client for interacting with the Synapse web services.
}
\details{
\tabular{ll}{
Package: \tab synapseClient\cr
Type: \tab Package\cr
Version: \tab 1.14-2\cr
Date: \tab 2017-01-09\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
}
}
\author{
Matt Furia and Nicole Deflaux
Maintainer: Bruce Hoff <bruce.hoff@sagebase.org>, Chris Bare <chris.bare@sagebase.org>
}
\keyword{ package }
\examples{
## Not Run

## End Not Run
}
