\name{synapseClient-package}
\alias{synapseClient-package}
\alias{synapseClient}
\docType{package}
\title{
	Synapse Client
}
\description{
	A client for interacting with the Synapse web services.
}
\details{
\tabular{ll}{
Package: \tab synapseClient\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-26\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}
}
\author{
	Matt Furia and Nicole Deflaux
	Maintainer: Matt Furia <matt.furia@sagebase.org>
}
\keyword{ package }
\examples{
	## Not Run
	
	## End Not Run
}
