\name{SynapseEntity}
\alias{Study}
\alias{Code}
\alias{Data}
\alias{SynapseEntity}
\alias{SynapseLocationOwner}
\alias{Project}
\alias{Analysis}
\alias{Step}
\title{
Entity Constructors
}
\description{
Constructors for Synapse Entities
}
\usage{
SynapseEntity(entity)
Study(entity)
Data(entity)
Project(entity)
}
\arguments{
  \item{entity}{
(optional) a named list whose elements will be set as the properties of the entity
}
}

\value{
The entity
}
\author{
Matt Furia
}
