\name{Used}
\alias{Used}
\docType{methods}
\title{
Used Constructor
}
\description{Constructor for Used}
\usage{Used(wasExecuted, concreteType)}
\arguments{
  \item{wasExecuted}{
  The enclosed entity was used and also executed in the Activity (logical)
  }
  \item{concreteType}{
   (character)
  }
}
\section{Slots}{
  \describe{
    \item{\code{wasExecuted}}{
    The enclosed entity was used and also executed in the Activity (logical)
    }
    \item{\code{concreteType}}{
     (character)
    }
  }
}
\seealso{
}
